增加pid精度

+++ Middlewares/Lw_Pid/lw_pid.c
@@ -10,13 +10,13 @@
 #define PID_MODE_FULLSCALE			4
 
 /*静态函数声明*/
-static float increment_pid_compute(Pid_object_t* pid_object, float current_input);
-static float fullscale_pid_compute(Pid_object_t* pid_object, float current_input);
+static double increment_pid_compute(Pid_object_t* pid_object, double current_input);
+static double fullscale_pid_compute(Pid_object_t* pid_object, double current_input);
 
 /*全局函数定义*/
 
 /*设置pid参数*/
-void pid_set_argument(Pid_object_t* pid_object, float kp, float ki, float kd)
+void pid_set_argument(Pid_object_t* pid_object, double kp, double ki, double kd)
 {
 	pid_object->kp = kp;
 	pid_object->ki = ki;
@@ -24,7 +24,7 @@ void pid_set_argument(Pid_object_t* pid_object, float kp, float ki, float kd)
 }
 
 /*设置pid目标值*/
-void pid_set_target(Pid_object_t* pid_object, float target)
+void pid_set_target(Pid_object_t* pid_object, double target)
 {
 	pid_object->target = target;
 }
@@ -42,7 +42,7 @@ void pid_output_limit_disable(Pid_object_t* pid_object)
 }
 
 /*设置pid输出限幅上下限*/
-void pid_set_output_limit(Pid_object_t* pid_object, float upper_limit, float floor_limit)
+void pid_set_output_limit(Pid_object_t* pid_object, double upper_limit, double floor_limit)
 {
 	pid_object->output_upper_limit = upper_limit;
 	pid_object->output_floor_limit = floor_limit;
@@ -61,7 +61,7 @@ void pid_intergral_limit_disable(Pid_object_t* pid_object)
 }
 
 /*设置pid的积分限幅上下限*/
-void pid_set_intergral_limit(Pid_object_t* pid_object, float upper_limit, float floor_limit)
+void pid_set_intergral_limit(Pid_object_t* pid_object, double upper_limit, double floor_limit)
 {
 	pid_object->intergral_upper_limit = upper_limit;
 	pid_object->intergral_floor_limit = floor_limit;
@@ -82,10 +82,10 @@ void pid_set_mode_fullscale(Pid_object_t* pid_object)
 /*初始化pid*/
 void pid_init(Pid_object_t* pid_object)
 {
-	pid_object->error0 = 0;
-	pid_object->error1 = 0;
-	pid_object->error2 = 0;
-	pid_object->error_intergral = 0;
+	pid_object->error0 = 0.0;
+	pid_object->error1 = 0.0;
+	pid_object->error2 = 0.0;
+	pid_object->error_intergral = 0.0;
 	
 	/*如果选择为增量式*/
 	if(pid_object->pid_mode == PID_MODE_INCREMENT)
@@ -101,9 +101,9 @@ void pid_init(Pid_object_t* pid_object)
 /*静态函数定义*/
 
 /*计算增量式pid*/
-static float increment_pid_compute(Pid_object_t* pid_object, float current_input)
+static double increment_pid_compute(Pid_object_t* pid_object, double current_input)
 {
-	float output;
+	double output;
 	
 	/*误差传递*/
 	pid_object->error2 = pid_object->error1;
@@ -130,9 +130,9 @@ static float increment_pid_compute(Pid_object_t* pid_object, float current_input
 }
 
 /*计算位置式pid*/
-static float fullscale_pid_compute(Pid_object_t* pid_object, float current_input)
+static double fullscale_pid_compute(Pid_object_t* pid_object, double current_input)
 {
-	float output;
+	double output;
 	
 	/*误差传递*/
 	pid_object->error1 = pid_object->error0;
